+++ MODSRC/vazkii/botania/client/gui/GuiLexiconIndex.java
@@ -15,11 +15,12 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import org.lwjgl.input.Mouse;
-
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.StatCollector;
+
+import org.lwjgl.input.Mouse;
+
 import vazkii.botania.api.lexicon.ILexicon;
 import vazkii.botania.api.lexicon.LexiconCategory;
 import vazkii.botania.api.lexicon.LexiconEntry;
@@ -87,7 +88,7 @@ public class GuiLexiconIndex extends GuiLexicon implements IParented {
 
 	@Override
 	protected void actionPerformed(GuiButton par1GuiButton) {
-		if(par1GuiButton.id >= BOOKMARK_START) 
+		if(par1GuiButton.id >= BOOKMARK_START)
 			handleBookmark(par1GuiButton);
 		else
 			switch(par1GuiButton.id) {
@@ -127,14 +128,14 @@ public class GuiLexiconIndex extends GuiLexicon implements IParented {
 	public void setParent(GuiLexicon gui) {
 		parent = gui;
 	}
-	
+
 	int fx = 0;
 	boolean swiped = false;
 
 	@Override
 	protected void mouseClickMove(int x, int y, int button, long time) {
 		if(button == 0 && Math.abs(x - fx) > 100 && mc.gameSettings.touchscreen && !swiped) {
-			double swipe = (double) (x - fx) / Math.max(1, (double) time);
+			double swipe = (x - fx) / Math.max(1, (double) time);
 			if(swipe < 0.5) {
 				nextPage();
 				swiped = true;
@@ -144,34 +145,34 @@ public class GuiLexiconIndex extends GuiLexicon implements IParented {
 			}
 		}
 	}
-	
+
 	@Override
 	protected void mouseClicked(int par1, int par2, int par3) {
 		super.mouseClicked(par1, par2, par3);
-		
+
 		fx = par1;
 		if(par3 == 1)
 			back();
 	}
-	
+
 	@Override
 	public void handleMouseInput() {
 		super.handleMouseInput();
-		
+
 		if(Mouse.getEventButton() == 0)
 			swiped = false;
-		
-        int w = Mouse.getEventDWheel();
-        if(w < 0)
-        	nextPage();
-        else if(w > 0)
-        	prevPage();
+
+		int w = Mouse.getEventDWheel();
+		if(w < 0)
+			nextPage();
+		else if(w > 0)
+			prevPage();
 	}
-	
+
 	@Override
 	protected void keyTyped(char par1, int par2) {
 		if(par2 == 203 || par2 == 200 || par2 == 201) // Left, Up, Page Up
-			prevPage();
+		prevPage();
 		else if(par2 == 205 || par2 == 208 || par2 == 209) // Right, Down Page Down
 			nextPage();
 		else if(par2 == 14) // Backspace
@@ -180,20 +181,20 @@ public class GuiLexiconIndex extends GuiLexicon implements IParented {
 			mc.displayGuiScreen(new GuiLexicon());
 			ClientTickHandler.notifyPageChange();
 		}
-		
+
 		super.keyTyped(par1, par2);
 	}
-	
+
 	void back() {
 		if(backButton.enabled)
 			actionPerformed(backButton);
 	}
-	
+
 	void nextPage() {
 		if(rightButton.enabled)
 			actionPerformed(rightButton);
 	}
-	
+
 	void prevPage() {
 		if(leftButton.enabled)
 			actionPerformed(leftButton);
