+++ MODSRC/vazkii/botania/client/render/entity/RenderPixie.java
@@ -53,7 +53,7 @@ public class RenderPixie extends RenderLiving {
 			char c0 = 61680;
 			int j = c0 % 65536;
 			int k = c0 / 65536;
-			OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, (float)j / 1.0F, (float)k / 1.0F);
+			OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit, j / 1.0F, k / 1.0F);
 			GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
 			GL11.glColor4f(1.0F, 1.0F, 1.0F, f1);
 			return 1;
